refactor: modify yml,separate notification
