+++ src/main/java/com/kt/notification/MailSendRequest.java
@@ -1,4 +1,4 @@
-package com.kt.dto.notify;
+package com.kt.notification;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
