+++ src/main/resources/application-local.yml
@@ -31,7 +31,7 @@ spring:
             enable: true
 
 jwt:
-  secret: ${kt.jwt.secret:kt-cloud-tech-up-shopping-202511171107}
+  secret: ${kt.jwt.secret}
   access-token-expiration: 300000
   refresh-token-expiration: 43200000
